Moved routing details into App.js, little more cleanup
