+++ src/app/index.js
@@ -1,4 +1,3 @@
 import App from './App'
-import store from './store'
 
-export {App, store}
+export default App
